Make helm upgrade redeploy all solutions
This is necessary if we're always using the tag 'latest', because
otherwise helm/kubernetes will just assume that latest == latest.

This seems to be a quirk with local development and not being able to
pull images from a central repository. Normally if you have a new tag or
tag latest kubernetes will always pull it seems.

An alternative solution is to just restart the deployment:
`kubectl rollout restart deployment <insert-deployment-name-here>`